rename vs code command and update
