+++ src/plugins/getPlugins.ts
@@ -1,5 +1,5 @@
 import { createAKSClusterPlugin } from "./createAKS/createAKSClusterPlugin";
-import { GetPluginsCommandResult } from "../types/@azure/AzureAgent";
+import { GetPluginsCommandResult } from "../types/aiazure/AzureAgent";
 import { generateKubectlCommandPlugin } from "./kubectlGeneration/generateKubectlCommandPlugin";
 
 export async function getPlugins(): Promise<GetPluginsCommandResult> {
