--- src/plugins/shared/constants.ts
@@ -1,1 +1,0 @@
-export const GITHUBCOPILOT_FOR_AZURE_VSCODE_ID = "ms-azuretools.vscode-azure-github-copilot"
